Refactor ProductType Typing into separate file
