+++ src/components/Product/ProductActions.tsx
@@ -8,7 +8,7 @@ import {
 import RemoveIcon from "@mui/icons-material/Remove";
 import AddIcon from "@mui/icons-material/Add";
 
-import { Product } from "../Products";
+import { Product } from "./ProductType";
 
 type ProductActionsType = {
   product: Product;
