+++ src/components/Product/ProductType.ts
@@ -1,0 +1,9 @@
+export type Product = {
+  id: number;
+  name: string;
+  imageUrl: string;
+  price: number;
+  category: string;
+  itemInCart: number;
+  loading: boolean;
+};
